+++ src/controllers/due.diligence.controller.ts
@@ -7,7 +7,7 @@ import {
 
 } from "../config";
 
-import { getDueDiligencePage, postDueDiligencePage } from "../utils/due-diligence";
+import { getDueDiligencePage, postDueDiligencePage } from "../utils/due.diligence";
 
 export const get = (req: Request, res: Response, next: NextFunction) => {
   getDueDiligencePage(req, res, next, DUE_DILIGENCE_PAGE, WHO_IS_MAKING_FILING_URL);
