+++ src/controllers/update/due.diligence.overseas.entity.controller.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 
-import { getDueDiligence, postDueDiligence } from "../../utils/due.diligence";
+import { getDueDiligence, postDueDiligence } from "../../utils/overseas.due.diligence";
 import {
   WHO_IS_MAKING_UPDATE_URL,
   UPDATE_DUE_DILIGENCE_OVERSEAS_ENTITY_PAGE,
@@ -8,7 +8,7 @@ import {
 } from "../../config";
 
 export const get = (req: Request, res: Response, next: NextFunction) => {
-  getDueDiligence(req, res, next, WHO_IS_MAKING_UPDATE_URL, UPDATE_DUE_DILIGENCE_OVERSEAS_ENTITY_PAGE);
+  getDueDiligence(req, res, next, UPDATE_DUE_DILIGENCE_OVERSEAS_ENTITY_PAGE, WHO_IS_MAKING_UPDATE_URL);
 };
 
 export const post = (req: Request, res: Response, next: NextFunction) => {
