+++ src/utils/due.diligence.ts
@@ -1,5 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import { Session } from "@companieshouse/node-session-handler";
+
 import { ApplicationData } from "../model";
 import { logger } from "../utils/logger";
 import {
@@ -9,27 +10,27 @@ import {
   mapDataObjectToFields,
   mapFieldsToDataObject,
 } from "../utils/application.data";
-import { DueDiligenceKey } from "../model/due.diligence.model";
+import { DueDiligenceKey, DueDiligenceKeys } from "../model/due.diligence.model";
+import { OverseasEntityDueDiligenceKey } from "../model/overseas.entity.due.diligence.model";
+import { IdentityDateKey, IdentityDateKeys } from "../model/date.model";
 import { IdentityAddressKey, IdentityAddressKeys } from "../model/address.model";
 import { AddressKeys, InputDateKeys } from "../model/data.types.model";
-import { IdentityDateKey, IdentityDateKeys } from "../model/date.model";
-import { OverseasEntityDueDiligenceKey, OverseasEntityDueDiligenceKeys } from "../model/overseas.entity.due.diligence.model";
-import { saveAndContinue } from "./save.and.continue";
+import { saveAndContinue } from "../utils/save.and.continue";
 
-export const getDueDiligence = (req: Request, res: Response, next: NextFunction, backLinkUrl: string, templateName: string) => {
+export const getDueDiligencePage = (req: Request, res: Response, next: NextFunction, templateName: string, backLinkUrl: string) => {
   try {
     logger.debugRequest(req, `${req.method} ${req.route.path}`);
 
     const appData: ApplicationData = getApplicationData(req.session);
-    const data = appData[OverseasEntityDueDiligenceKey];
+    const agentData = appData[DueDiligenceKey];
 
-    const identityAddress = (data?.[IdentityAddressKey]) ? mapDataObjectToFields(data[IdentityAddressKey], IdentityAddressKeys, AddressKeys) : {};
-    const identityDate = (data?.[IdentityDateKey]) ? mapDataObjectToFields(data[IdentityDateKey], IdentityDateKeys, InputDateKeys) : {};
+    const identityAddress = (agentData?.[IdentityAddressKey]) ? mapDataObjectToFields(agentData[IdentityAddressKey], IdentityAddressKeys, AddressKeys) : {};
+    const identityDate = (agentData?.[IdentityDateKey]) ? mapDataObjectToFields(agentData[IdentityDateKey], IdentityDateKeys, InputDateKeys) : {};
 
     return res.render(templateName, {
       backLinkUrl: backLinkUrl,
       templateName: templateName,
-      ...data,
+      ...agentData,
       ...identityAddress,
       [IdentityDateKey]: identityDate
     });
@@ -38,19 +39,19 @@ export const getDueDiligence = (req: Request, res: Response, next: NextFunction,
   }
 };
 
-export const postDueDiligence = async (req: Request, res: Response, next: NextFunction, redirectUrl: string, registrationFlag: boolean): Promise<void> => {
+export const postDueDiligencePage = async (req: Request, res: Response, next: NextFunction, redirectUrl: string, registrationFlag: boolean): Promise<void> => {
   try {
     logger.debugRequest(req, `${req.method} ${req.route.path}`);
 
     const session = req.session as Session;
-    const data = prepareData(req.body, OverseasEntityDueDiligenceKeys);
-    data[IdentityAddressKey] = mapFieldsToDataObject(req.body, IdentityAddressKeys, AddressKeys);
-    data[IdentityDateKey] = mapFieldsToDataObject(req.body, IdentityDateKeys, InputDateKeys);
+    const agentData = prepareData(req.body, DueDiligenceKeys);
+    agentData[IdentityAddressKey] = mapFieldsToDataObject(req.body, IdentityAddressKeys, AddressKeys);
+    agentData[IdentityDateKey] = mapFieldsToDataObject(req.body, IdentityDateKeys, InputDateKeys);
 
-    setApplicationData(session, data, OverseasEntityDueDiligenceKey);
+    setApplicationData(session, agentData, DueDiligenceKey);
 
-    // Empty DueDiligence object
-    setApplicationData(session, {}, DueDiligenceKey);
+    // Empty OverseasEntityDueDiligence object
+    setApplicationData(session, {}, OverseasEntityDueDiligenceKey);
 
     if (registrationFlag) {
       await saveAndContinue(req, session);
